removed "Empty" variant of "Mark"
it is un-idiomatic to use an "Empty" enum variant in rust. instead, it is more desirable to only give your enum *valid* values it can be.

for situations where there is *no* mark, we wrap the mark in an Option, so we end up with Option<Mark>.

now, every tile can be one of 3 things:

None
Some(Circle)
Some(Cross)

it's subtle why this is better, but an example of where this makes the code nicer is for game's "turn" variable.

when playing, the only two *valid* values for "turn" are Cross and Circle... but with the previous variant, "Empty" was one of the problems.

now, that issue could never possibly happen in our code, because rust will never allow "turn" to be that value because we've removed it completely.