iterator-based collection
your code is just fine here, but i'm just making this change to demonstrate an alternative approach to writing functions like this that "collect" values.

here, we "enumerate" the board values and map each tile like so:

Some => None
None => cord of tile

the "filter" part of the filter_map() then discards all the "None" values (so we're left with just an iterator of the tile numbers that were empty).

then, we can collect() that iterator into a vector and return it.

the choice is yours, both your method and this method are valid, but the more i write rust, the more i start to love how the functional-style iterators look.